docs(2020): add documentation to 2020 day 1

+++ rust/2020/src/day1.rs
@@ -1,7 +1,11 @@
 use std::collections::HashSet;
 
+// Day 1: Report Repair
+// A version of "Two Sum Problem"
+// https://leetcode.com/problems/two-sum/solution/
+
 #[aoc_generator(day1)]
-pub fn input_generator(input: &str) -> HashSet<u64> {
+pub fn input_generator(input: &str) -> HashSet<u32> {
     input
         .lines()
         .map(|l| l.trim())
@@ -10,8 +14,35 @@ pub fn input_generator(input: &str) -> HashSet<u64> {
         .collect()
 }
 
+/*
+ * Before you leave, the Elves in accounting just need you to fix your
+ * expense report (your puzzle input); apparently, something isn't quite adding up.
+ *
+ * Specifically, they need you to find the two entries that sum to
+ * `2020` and then multiply those two numbers together.
+ *
+ * For example, suppose your expense report contained the following:
+ *
+ * 1721
+ * 979
+ * 366
+ * 299
+ * 675
+ * 1456
+ *
+ * In this list, the two entries that sum to 2020 are 1721 and 299.
+ * Multiplying them together produces 1721 * 299 = 514579, so the correct answer is 514579.
+ *
+ * Of course, your expense report is much larger. Find the two entries that sum to 2020; what do you get if you multiply them together?
+*/
+/// Your puzzle answer was
+/// ```
+/// use advent_of_code_2020::day1::*;
+/// let data = include_str!("../input/2020/day1.txt");
+/// assert_eq!(solve_part_01(&input_generator(data)), 898299);
+/// ```
 #[aoc(day1, part1)]
-pub fn solve_part_01(input: &HashSet<u64>) -> u64 {
+pub fn solve_part_01(input: &HashSet<u32>) -> u32 {
     for x in input {
         let y = 2020 - x;
         if input.contains(&y) {
@@ -22,8 +53,24 @@ pub fn solve_part_01(input: &HashSet<u64>) -> u64 {
     0
 }
 
+/*
+ * The Elves in accounting are thankful for your help; one of them even offers you
+ * a starfish coin they had left over from a past vacation. They offer you a second
+ * one if you can find three numbers in your expense report that meet the same criteria.
+ *
+ * Using the above example again, the three entries that sum to 2020 are 979, 366, and 675.
+ * Multiplying them together produces the answer, 241861950.
+ *
+ * In your expense report, what is the product of the three entries that sum to 2020?
+*/
+/// Your puzzle answer was
+/// ```
+/// use advent_of_code_2020::day1::*;
+/// let data = include_str!("../input/2020/day1.txt");
+/// assert_eq!(solve_part_02(&input_generator(data)), 143933922);
+/// ```
 #[aoc(day1, part2)]
-pub fn solve_part_02(input: &HashSet<u64>) -> u64 {
+pub fn solve_part_02(input: &HashSet<u32>) -> u32 {
     for x in input {
         for y in input {
             if x + y > 2020 {
